Port: handle 'size_t' / 'off_t' on Solaris
The size of 'size_t' / 'off_t' vary on some platforms, use C99 portable
macros and print formats to handle them.

Print 'size_t' with '%zu' and use SIZE_MAX as max value.
Print 'off_t' casted to 'intmax_t' with '%jd'.

+++ plugins/ctf/fs-src/file.c
@@ -101,7 +101,7 @@ int ctf_fs_file_open(struct ctf_fs_file *file, const char *mode)
 	}
 
 	file->size = stat.st_size;
-	BT_LOGD("File is %zu bytes", file->size);
+	BT_LOGD("File is %jd bytes", (intmax_t) file->size);
 	goto end;
 
 error:
